# perf: optimize URL serialization