--- a/core/Cargo.toml
+++ b/core/Cargo.toml
@@ -32,7 +32,7 @@ serde = { version = "1.0.136", features = ["derive"] }
 serde_json = { version = "1.0.79", features = ["preserve_order"] }
 serde_v8 = { version = "0.60.0", path = "../serde_v8" }
 sourcemap = "=6.0.1"
-url = { version = "2.3.0", features = ["serde"] }
+url = { version = "2.3.0", features = ["serde", "expose_internals"] }
 v8 = { version = "0.49.0", default-features = false }
 
 [[example]]