--- a/ext/url/00_url.js
+++ b/ext/url/00_url.js
@@ -21,7 +21,6 @@
     ObjectKeys,
     SafeArrayIterator,
     StringPrototypeSlice,
-    StringPrototypeSplit,
     Symbol,
     SymbolFor,
     SymbolIterator,
@@ -44,41 +43,38 @@
 
   // Helper functions
   function opUrlReparse(href, setter, value) {
-    return _urlParts(
-      ops.op_url_reparse(href, [setter, value]),
-    );
+    if (!componentsBuf) {
+      componentsBuf = new Uint32Array(8);
+      core.ops.op_url_set_buf(componentsBuf);
+    }
+
+    const status = ops.op_url_reparse(href, setter, value);
+    return getSerialization(status, href);
   }
+
   function opUrlParse(href, maybeBase) {
-    return _urlParts(ops.op_url_parse(href, maybeBase));
+    if (!componentsBuf) {
+      componentsBuf = new Uint32Array(8);
+      core.ops.op_url_set_buf(componentsBuf);
+    }
+
+    let status;
+    if (maybeBase === undefined) {
+      status = ops.op_url_parse(href);
+    } else {
+      status = core.ops.op_url_parse_with_base(href, maybeBase);
+    }
+    return getSerialization(status, href);
   }
-  function _urlParts(internalParts) {
-    // WARNING: must match UrlParts serialization rust's url_result()
-    const {
-      0: href,
-      1: hash,
-      2: host,
-      3: hostname,
-      4: origin,
-      5: password,
-      6: pathname,
-      7: port,
-      8: protocol,
-      9: search,
-      10: username,
-    } = StringPrototypeSplit(internalParts, "\n");
-    return {
-      href,
-      hash,
-      host,
-      hostname,
-      origin,
-      password,
-      pathname,
-      port,
-      protocol,
-      search,
-      username,
-    };
+
+  function getSerialization(status, href) {
+    if (status === 0) {
+      return href;
+    } else if (status === 1) {
+      return core.ops.op_url_get_serialization();
+    } else {
+      throw new TypeError("Invalid URL");
+    }
   }
 
   class URLSearchParams {
@@ -131,7 +127,7 @@
       if (url === null) {
         return;
       }
-      url[_url] = opUrlReparse(url.href, SET_SEARCH, this.toString());
+      url[updateUrlSearch](this.toString());
     }
 
     /**
@@ -308,11 +304,25 @@
     URLSearchParamsPrototype,
   );
 
-  const _url = Symbol("url");
+  const _updateUrlSearch = Symbol("updateUrlSearch");
 
+  let componentsBuf;
   class URL {
-    [_url];
     #queryObject = null;
+    #serialization;
+    #schemeEnd;
+    #usernameEnd;
+    #hostStart;
+    #hostEnd;
+    #port;
+    #pathStart;
+    #queryStart;
+    #fragmentStart;
+
+    [_updateUrlSearch](value) {
+      opUrlReparse(this.#serialization, SET_SEARCH, value);
+      this.#updateComponents();
+    }
 
     /**
      * @param {string} url
@@ -328,7 +338,21 @@
         });
       }
       this[webidl.brand] = webidl.brand;
-      this[_url] = opUrlParse(url, base);
+      this.#serialization = opUrlParse(url, base);
+      this.#updateComponents();
+    }
+
+    #updateComponents() {
+      [
+        this.#schemeEnd,
+        this.#usernameEnd,
+        this.#hostStart,
+        this.#hostEnd,
+        this.#port,
+        this.#pathStart,
+        this.#queryStart,
+        this.#fragmentStart,
+      ] = componentsBuf;
     }
 
     [SymbolFor("Deno.privateCustomInspect")](inspect, inspectOptions) {
@@ -366,7 +390,10 @@
     /** @return {string} */
     get hash() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].hash;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/quirks.rs#L263
+      return this.#fragmentStart
+        ? this.#serialization.slice(this.#fragmentStart)
+        : "";
     }
 
     /** @param {string} value */
@@ -379,7 +406,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_HASH, value);
+        opUrlReparse(this.#serialization, SET_HASH, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -388,7 +416,8 @@
     /** @return {string} */
     get host() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].host;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/quirks.rs#L101
+      return this.#serialization.slice(this.#hostStart, this.#pathStart);
     }
 
     /** @param {string} value */
@@ -401,7 +430,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_HOST, value);
+        opUrlReparse(this.#serialization, SET_HOST, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -410,7 +440,8 @@
     /** @return {string} */
     get hostname() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].hostname;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/lib.rs#L988
+      return this.#serialization.slice(this.#hostStart, this.#hostEnd);
     }
 
     /** @param {string} value */
@@ -423,7 +454,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_HOSTNAME, value);
+        opUrlReparse(this.#serialization, SET_HOSTNAME, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -432,7 +464,7 @@
     /** @return {string} */
     get href() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].href;
+      return this.#serialization;
     }
 
     /** @param {string} value */
@@ -444,20 +476,59 @@
         prefix,
         context: "Argument 1",
       });
-      this[_url] = opUrlParse(value);
+      this.#serialization = opUrlParse(value);
+      this.#updateComponents();
       this.#updateSearchParams();
     }
 
     /** @return {string} */
     get origin() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].origin;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/origin.rs#L14
+      const scheme = this.#serialization.slice(0, this.#schemeEnd);
+      if (
+        scheme === "http" || scheme === "https" || scheme === "ftp" ||
+        scheme === "ws" || scheme === "wss"
+      ) {
+        if (this.port) {
+          return `${scheme}://${this.host}/${this.port}`;
+        } else {
+          return `${scheme}://${this.host}`;
+        }
+      }
+
+      if (scheme === "blob") {
+        // TODO(@littledivy): Fast path.
+        try {
+          return new URL(this.pathname).origin;
+        } catch {
+          return "null";
+        }
+      }
+
+      return "null";
     }
 
     /** @return {string} */
     get password() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].password;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/lib.rs#L914
+      if (
+        this.#hasAuthority() &&
+        this.#usernameEnd !== this.#serialization.length &&
+        this.#serialization[this.#usernameEnd] === ":"
+      ) {
+        return this.#serialization.slice(
+          this.#usernameEnd + 1,
+          this.#hostStart - 1,
+        );
+      }
+      return "";
+    }
+
+    #hasAuthority() {
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/lib.rs#L824
+      return this.#serialization.slice(this.#schemeEnd).startsWith("://");
     }
 
     /** @param {string} value */
@@ -470,7 +541,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_PASSWORD, value);
+        opUrlReparse(this.#serialization, SET_PASSWORD, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -479,7 +551,13 @@
     /** @return {string} */
     get pathname() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].pathname;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/lib.rs#L1203
+      if (!this.#queryStart && !this.#fragmentStart) {
+        return this.#serialization.slice(this.#pathStart);
+      }
+
+      const nextComponentStart = this.#queryStart || this.#fragmentStart;
+      return this.#serialization.slice(this.#pathStart, nextComponentStart);
     }
 
     /** @param {string} value */
@@ -492,7 +570,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_PATHNAME, value);
+        opUrlReparse(this.#serialization, SET_PATHNAME, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -501,7 +580,15 @@
     /** @return {string} */
     get port() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].port;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/quirks.rs#L196
+      if (this.#port) {
+        return this.#serialization.slice(
+          this.#hostEnd + 1, /* : */
+          this.#pathStart,
+        );
+      } else {
+        return this.#serialization.slice(this.#hostEnd, this.#pathStart);
+      }
     }
 
     /** @param {string} value */
@@ -514,7 +601,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_PORT, value);
+        opUrlReparse(this.#serialization, SET_PORT, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -523,7 +611,8 @@
     /** @return {string} */
     get protocol() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].protocol;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/quirks.rs#L56
+      return this.#serialization.slice(0, this.#schemeEnd + 1 /* : */);
     }
 
     /** @param {string} value */
@@ -536,7 +625,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_PROTOCOL, value);
+        opUrlReparse(this.#serialization, SET_PROTOCOL, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -545,7 +635,11 @@
     /** @return {string} */
     get search() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].search;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/quirks.rs#L249
+      const afterPath = this.#queryStart || this.#fragmentStart ||
+        this.#serialization.length;
+      const afterQuery = this.#fragmentStart || this.#serialization.length;
+      return this.#serialization.slice(afterPath, afterQuery);
     }
 
     /** @param {string} value */
@@ -558,7 +652,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_SEARCH, value);
+        opUrlReparse(this.#serialization, SET_SEARCH, value);
+        this.#updateComponents();
         this.#updateSearchParams();
       } catch {
         /* pass */
@@ -568,7 +663,19 @@
     /** @return {string} */
     get username() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].username;
+      // https://github.com/servo/rust-url/blob/1d307ae51a28fecc630ecec03380788bfb03a643/url/src/lib.rs#L881
+      const schemeSeperatorLen = 3; /* :// */
+      if (
+        this.#hasAuthority() &&
+        this.#usernameEnd > this.#schemeEnd + schemeSeperatorLen
+      ) {
+        return this.#serialization.slice(
+          this.#schemeEnd + schemeSeperatorLen,
+          this.#usernameEnd,
+        );
+      } else {
+        return "";
+      }
     }
 
     /** @param {string} value */
@@ -581,7 +688,8 @@
         context: "Argument 1",
       });
       try {
-        this[_url] = opUrlReparse(this[_url].href, SET_USERNAME, value);
+        opUrlReparse(this.#serialization, SET_USERNAME, value);
+        this.#updateComponents();
       } catch {
         /* pass */
       }
@@ -599,13 +707,13 @@
     /** @return {string} */
     toString() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].href;
+      return this.#serialization;
     }
 
     /** @return {string} */
     toJSON() {
       webidl.assertBranded(this, URLPrototype);
-      return this[_url].href;
+      return this.#serialization;
     }
   }