2.4x improvement in `(new URL(path)).pathname`.

```
# main
deno run -A --unstable cli/bench/url_parse.js 3 1000000
time 1080 ms rate 925925
time 1064 ms rate 939849
time 1048 ms rate 954198

# this patch
target/release/deno run -A --unstable cli/bench/url_parse.js 3 1000000
time 444 ms rate 2252252
time 429 ms rate 2331002
time 474 ms rate 2109704
```